Revert "Remove uneeded type conversion"
This reverts commit efcf97d624aca1df50469a649ede6813dc1e7623. At 32bit platforms (e.g. armv6h)
it causes a compile time error:
  ./purge.go:43:31: cannot use atimeUnix.Sec (type int32) as type int64 in argument to time.Unix
  ./purge.go:43:46: cannot use atimeUnix.Nsec (type int32) as type int64 in argument to time.Unix

Closes #18

+++ purge.go
@@ -40,7 +40,8 @@ func purgeStaleFiles(cacheDir string, purgeFilesAfter int) {
 		}
 
 		atimeUnix := info.Sys().(*syscall.Stat_t).Atim
-		atime := time.Unix(atimeUnix.Sec, atimeUnix.Nsec)
+		// Note that int64() is needed here, otherwise compilation fails at 32 bit platforms like armv6h. See issue #18.
+		atime := time.Unix(int64(atimeUnix.Sec), int64(atimeUnix.Nsec))
 		if atime.Before(removeIfOlder) {
 			log.Printf("Remove stale file %v as its access time (%v) is too old", path, atime)
 			err := os.Remove(path)
